+++ spec/rails_helper.rb
@@ -50,6 +50,7 @@ RSpec.configure do |config|
   config.include TransactionHelpers
   config.include ExportHelpers
   config.include Devise::Test::ControllerHelpers, type: :controller
+  config.include CsvImportHelpers, type: :import_csv
 
   # Remove this line if you're not using ActiveRecord or ActiveRecord fixtures
   config.fixture_path = "#{::Rails.root}/spec/fixtures"